LinBPQ 6.0.21
Single commit of the released version as available on the .zip at
https://www.cantab.net/users/john.wiseman/Downloads/ .

At this moment in time, there is already a BPQ32 6.0.22, but not the
correspondent LinBPQ (at least no code drop).

+++ Moncode.c
@@ -197,7 +197,7 @@ KC6OAR*>ID:
 
 		Output += sprintf((char *)Output, " Port=%d\r", Port);
 
-		return strlen(buffer);
+		return (int)strlen(buffer);
 	}
 
 	n = 8;						// MAX DIGIS
